ip proxy fix after lab testing.

+++ dnx_iptools/dnx_trie_search/dnx_trie_search.pxd
@@ -59,4 +59,4 @@ cdef class RangeTrie:
 
     cpdef void generate_structure(self, tuple py_trie)
     cdef uint32_t search(self, uint32_t container_id, uint32_t host_id) nogil
-    cdef L2Range* make_l2(self, (long, long, short) l2_entry)
+    cdef L2Range* make_l2(self, (uint32_t, uint32_t, uint16_t) l2_entry)
